test(runtimes): fold the continue-gate tests into the file that mirrors their source
`test__tui_continue_gate.py` had no mirroring source module, so PS-204 §2
(orphan-test-file) failed `tests/develop/test_audit.py::test_audit_all_clean`
on every pytest-matrix leg:

    [E] [PS-204 §2 orphan-test-file] ...: no matching src file (orphan test);
        mirror dir `src/scitex_agent_container/runtimes` contains: ... none
        match expected `_tui_continue_gate.py`

The gate itself is not a module of its own -- `_home_has_resumable_conversation`
lives in `_apptainer_inner_argv_tui.py`, which is where the tests import it
from and which already has a mirroring test file. So the tests belong there,
not behind a new name; renaming the file would have produced a second orphan
rather than a fix (measured: `test__apptainer_inner_argv_tui_continue_gate.py`
has no mirroring source either).

The measured story the file carried is preserved verbatim as a comment block
above the folded-in cases, because it is the reason the gate is being widened:
on the 2026-08-11 canary relocation to scitex-compute-04 the transcript arrived
byte- and line-verified, the marker was seeded, and the agent still started
FRESH -- the gate probed the overlay upper-home (which existed and was empty)
and never the bind carrying the transcript, so `-c` was silently omitted and
3.7 MB of conversation one directory away was not resumed.

Measured: 8 passed, rc 0.
